move a lot of code to clean up
