+++ src/pages/discover/discoverContent.tsx
@@ -315,84 +315,10 @@ export function DiscoverContent() {
   };
 
   const handleShowDetails = async (media: MediaItem) => {
-    setIsLoadingDetails(true);
-    try {
-      const type =
-        media.type === "movie" ? TMDBContentTypes.MOVIE : TMDBContentTypes.TV;
-      const details = await getMediaDetails(media.id, type);
-      const backdropUrl = getMediaBackdrop(details.backdrop_path);
-
-      if (type === TMDBContentTypes.MOVIE) {
-        const movieDetails = details as TMDBMovieData;
-        setDetailsData({
-          title: movieDetails.title,
-          overview: movieDetails.overview,
-          backdrop: backdropUrl,
-          runtime: movieDetails.runtime,
-          genres: movieDetails.genres,
-          language: movieDetails.original_language,
-          voteAverage: movieDetails.vote_average,
-          voteCount: movieDetails.vote_count,
-          releaseDate: movieDetails.release_date,
-          rating: movieDetails.release_dates?.results?.find(
-            (r) => r.iso_3166_1 === "US",
-          )?.release_dates?.[0]?.certification,
-          director: movieDetails.credits?.crew?.find(
-            (person) => person.job === "Director",
-          )?.name,
-          actors: movieDetails.credits?.cast
-            ?.slice(0, 5)
-            .map((actor) => actor.name),
-          type: "movie",
-          id: movieDetails.id,
-          imdbId: movieDetails.external_ids?.imdb_id,
-        });
-      } else {
-        const showDetails = details as TMDBShowData & {
-          episodes: Array<{
-            id: number;
-            name: string;
-            episode_number: number;
-            overview: string;
-            still_path: string | null;
-            air_date: string;
-            season_number: number;
-          }>;
-        };
-        setDetailsData({
-          title: showDetails.name,
-          overview: showDetails.overview,
-          backdrop: backdropUrl,
-          episodes: showDetails.number_of_episodes,
-          seasons: showDetails.number_of_seasons,
-          genres: showDetails.genres,
-          language: showDetails.original_language,
-          voteAverage: showDetails.vote_average,
-          voteCount: showDetails.vote_count,
-          releaseDate: showDetails.first_air_date,
-          rating: showDetails.content_ratings?.results?.find(
-            (r) => r.iso_3166_1 === "US",
-          )?.rating,
-          director: showDetails.credits?.crew?.find(
-            (person) => person.job === "Director",
-          )?.name,
-          actors: showDetails.credits?.cast
-            ?.slice(0, 5)
-            .map((actor) => actor.name),
-          type: "show",
-          id: showDetails.id,
-          imdbId: showDetails.external_ids?.imdb_id,
-          seasonData: {
-            seasons: showDetails.seasons,
-            episodes: showDetails.episodes,
-          },
-        });
-      }
-    } catch (err) {
-      console.error("Failed to fetch media details:", err);
-    } finally {
-      setIsLoadingDetails(false);
-    }
+    setDetailsData({
+      id: Number(media.id),
+      type: media.type === "movie" ? "movie" : "show",
+    });
     detailsModal.show();
   };
 
@@ -591,11 +517,7 @@ export function DiscoverContent() {
 
       <ScrollToTopButton />
 
-      <DetailsModal
-        id="discover-details"
-        data={detailsData}
-        isLoading={isLoadingDetails}
-      />
+      {detailsData && <DetailsModal id="discover-details" data={detailsData} />}
     </div>
   );
 }
